hardware_gpio: use GPIO0.IE workaround for Errata E9
When a gpio on the rp2350 is not pulled up it can have the gpio latching
problem. One way to solve this is by clearing GPIO0.IE according to the
Errata E9. To read from the GPIO the GPIO0.IE bits have to be set before
reading and cleared afterwards.

Resolves issue #8

Signed-off-by: bauer-alex-174 <bauer.alex.174@gmail.com>